+++ source/FlexTable/index.js
@@ -1,5 +1,6 @@
 /* @flow */
 export default from './FlexTable'
+export defaultRowRenderer from './defaultRowRenderer'
 export FlexTable from './FlexTable'
 export FlexColumn from './FlexColumn'
 export SortDirection from './SortDirection'